Refactor: Extract Edit Expenditure component
