--- src/components/common/Expenditures/EditLockedExpenditure/index.ts
@@ -1,1 +1,0 @@
-export { default } from './EditLockedExpenditureButton';
