+++ src/components/common/Expenditures/ExpenditureForm/EditExpenditure/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EditExpenditure';
