+++ src/components/common/Expenditures/ExpenditureForm/index.ts
@@ -1,4 +1,4 @@
-export { default as EditExpenditureForm } from './EditExpenditureForm';
+export { default as EditExpenditure } from './EditExpenditure';
 export { default as CreateExpenditure } from './CreateExpenditure';
 export * from './types';
 export * from './helpers';
